WIP
Add Type mapping and test
