+++ src/main/scala/dpla/ingestion3/enrichments/EnrichmentDriver.scala
@@ -22,6 +22,7 @@ class EnrichmentDriver(conf: i3Conf) extends Serializable {
   val dateEnrichment = new ParseDateEnrichment()
   val spatialEnrichment = new SpatialEnrichment(Geocoder)
   val languageMapper = new LanguageMapper
+  val typeMapper = new TypeMapper
 
   /**
     * Applies a set of common enrichments that need to be run for all providers
@@ -43,22 +44,8 @@ class EnrichmentDriver(conf: i3Conf) extends Serializable {
       sourceResource = enriched.sourceResource.copy(
         date = enriched.sourceResource.date.map(d => dateEnrichment.parse(d)),
         language = enriched.sourceResource.language.map(l => languageMapper.enrichLanguage(l)),
-        place = enriched.sourceResource.place.map(p => spatialEnrichment.enrich(p)),
-
-        /**
-          Type enrichment
-          ----------------
-          Lower case the original string value and try to map to DCMIType IRIs (@see mapDcmiType()).
-          If the original type value can be mapped to a valid IRI then mapDcmiTypeToString() maps the
-          IRI to a string label and lowercases that label. The lowercase IRI label (without the corresponding
-          IRI) is the enriched value.
-
-          Original values that cannot be mapped to a DCMIType IRI are dropped.
-         */
-        `type` = enriched.sourceResource.`type`
-          .flatMap(t => DcmiTypeMapper.mapDcmiType(t.toLowerCase()))
-          .map(DcmiTypeStringMapper.mapDcmiTypeString)
-          .map(_.toLowerCase)
+        // place = enriched.sourceResource.place.map(p => spatialEnrichment.enrich(p)),
+        `type` = enriched.sourceResource.`type`.flatMap(t => typeMapper.enrich(t))
         )
       )
   }
