+++ src/main/scala/dpla/ingestion3/enrichments/LanguageMapper.scala
@@ -6,43 +6,64 @@ import dpla.ingestion3.utils.FileLoader
 /**
   * Language mapping examples
   */
-class LanguageMapper extends FileLoader with VocabMapper[SkosConcept] {
+class LanguageMapper extends FileLoader with VocabMapper[SkosConcept] with VocabValidator[SkosConcept] {
 
-  private val mapperLookup = new MapperLookup[SkosConcept] (
-    // Normalizations terms for retrieval
-    (term: SkosConcept) => term
-      .providedLabel.getOrElse("")
-      .toLowerCase()
-      .trim,
-    // Merges enriched and original values
-    (prov: SkosConcept, enriched: SkosConcept) => enriched.copy(providedLabel = prov.providedLabel),
-    // Validates provided term against controlled vocab terms
-    (prov: SkosConcept, enriched: SkosConcept) => prov
-      .providedLabel.getOrElse("")
-      .trim
-      .equalsIgnoreCase(enriched.concept.get.trim)
-  )
-
-  // Load vocab
-  loadVocab
-
-  // FileLoader
-  override def files: Seq[String] = Seq(
+  private val fileList = Seq(
     "/languages/iso639-2.csv",
     "/languages/iso639-3.csv",
     "/languages/dpla-lang.csv"
   )
 
+
+  private val mapperLookup = new MapperLookup[SkosConcept](
+    (term: SkosConcept) => normalizationFunc(term),
+    (prov: SkosConcept, enriched: SkosConcept) => mergeFunc(prov, enriched)
+  )
+
+  private val mapperValidate = new MapperLookup[SkosConcept](
+    (term: SkosConcept) => normalizationFunc(term),
+    (prov: SkosConcept, enriched: SkosConcept) => mergeFunc(prov, enriched)
+  )
+
+  private def normalizationFunc(term: SkosConcept): String = term
+      .providedLabel.getOrElse("")
+      .toLowerCase()
+      .trim
+
+  private def mergeFunc(prov: SkosConcept, enriched: SkosConcept) =
+    enriched.copy(providedLabel = prov.providedLabel)
+
   //noinspection TypeAnnotation
-  private def loadVocab = {
-    getVocabFromCsvFiles(files)
-      .map(term => {
-        mapperLookup.add(
-          SkosConcept(
-            providedLabel = Some(term(0)),
-            concept = Some(term(1)))
-        )
-      })
+  private def loadVocab =
+    getVocabFromCsvFiles(files).map(term => addLangConcept(term(0), term(1)))
+
+  // Load vocab
+  loadVocab
+
+  //noinspection TypeAnnotation
+  private def addLangConcept(provLabel: String, concept: String) = {
+    // Use term abbrev for lookup key
+    mapperLookup.add(
+      SkosConcept(
+        providedLabel = Some(provLabel),
+        concept = Some(concept)
+    ))
+    // Use full term for lookup key
+    mapperValidate.add(
+      SkosConcept(
+        providedLabel = Some(concept),
+        concept = Some(concept)
+    ))
+  }
+
+  // FileLoader
+  override def files: Seq[String] = fileList
+
+  override def enrich(value: SkosConcept): Option[SkosConcept] =
+    mapperLookup.lookup(value)
+
+  override def validate(value: SkosConcept): Option[SkosConcept] = {
+    mapperValidate.lookup(value)
   }
 
   // Does both abv > term enrichment and validation of original value as matching controlled vocab
@@ -50,17 +71,10 @@ class LanguageMapper extends FileLoader with VocabMapper[SkosConcept] {
     (enrich(value), validate(value)) match {
       case (Some(e), _) => mapperLookup.merge(value, e)
       case (None, Some(v)) => mapperLookup.merge(value, v)
-      case (None, None) => value
+      case (_, _) => value
     }
   }
-  override def enrich(value: SkosConcept): Option[SkosConcept] = {
-    mapperLookup
-      .lookup(value)
-  }
 
-  override def validate(value: SkosConcept): Option[SkosConcept] = {
-    mapperLookup
-      .validate(value)
-  }
+
 }
